Update withProperties's handling of updated text.
Because we deal with text more or less like HTML -- as child
nodes -- we need to deal with various situations where text
is specified and there are other child nodes.

Currently we deal with this by combining updated text and
children by prepending the text to the incoming children. If
there are no updated chidlren but there are existing ones,
the updated text replaces any leading text in the existing
children. Lastly, if there are no existing or updated children,
the element ends up with a single node that is the incoming
text.