Add rerendering pipeline, attached to the element-changed event.
The element-changed event is subscribed to using jquery,
and it’s expected to have a before and an after property with
the GraftElement instance from before the change and the
one after the change.

Rerender requests update the htmlized roots so that when
the rerender actually runs, it is htmlizing the correct roots.

Rerenders are scheduled using requestAnimationFrame,
which degrades to a setTimeout when the standard requestAnimationFrame
is not available.